Update the root view
Update to include a success flash for saving the wishlist.

Include a section to tell you who you have to buy a secret santa for as
a placeholder.

Apply some styling, headings and spacing and include a new panel css to
display the wishlist in because it looks fancy.

Display your own wishlist and turn it in to basic html with
`simple_format`