Add secret santa model and create relationships
The secret santa model is a self join table for users.

Set up the relationships as well so we expose convenience methods on
`User` that will allow us to query their `#secret_santas` and
`#recipients`.

We set the following on `User`:

```
  has_many :user_to_user_as_santa, foreign_key: "santa_id", class_name: "SecretSanta"
  has_many :recipients, through: :user_to_user_as_santa, source: :recipient

  has_many :user_to_user_as_recipient, foreign_key: "recipient_id", class_name: "SecretSanta"
  has_many :secret_santas, through: :user_to_user_as_recipient, source: :santa
```

This can seem confusing so it is worth some comment.

`SecretSanta` defines a many to many relationship for `User` to `User`.
To do so, it defines 2 fields. `santa_id` and `recipient_id`. So for a
single row in `SecretSanta`, we model a relatonship of secret santa to
gift recipient. Both fields are integers and foreign keys to the users
table (as modelled by `User`)

For a `User` instance, if we wanted to find all people for whom they
should buy a present, we could query `SecretSanta` where `santa_id`
equals the instance's `User#id`. Our SQL might look something like:

```
select recipient_id from secret_santas where santa_id = ?
```

where `?` represents the current user id. To enhance this we can join
to the users table and return the full details of the recipients.

```
select name, wishlist
from users inner join secret_santas
  on users.id = secret_santas.recipient_id
where secret_santas.santa_id = ?
```

To expose some convenience methods on `User` instances so we do not have
to write the SQL, we set up the four `has_many` relationships.

```
  has_many :user_to_user_as_santa, foreign_key: "santa_id", class_name: "SecretSanta"
  has_many :recipients, through: :user_to_user_as_santa, source: :recipient

  has_many :user_to_user_as_recipient, foreign_key: "recipient_id", class_name: "SecretSanta"
  has_many :secret_santas, through: :user_to_user_as_recipient, source: :santa
```

The `user_to_user_as_santa` will allow us to retrieve all rows in
`secret_santas` where the current user is the secret santa. The second
relationship `has_many :recipients, through: user_to_user_as_santa`
might seem unintuitive at first, but it allows Rails to replicate the
the second SQL statement above. Thus it returns all rows in
`secret_santas` where the current user id is the same as `santa_id`.
This means all the recordset of `recipient_id` returned represents all
the people for whom the current user must buy a secret santa present.

We continue to define `source: :recipient` which references the named
`belongs_to` definition in `SecretSanta`.

So, we can think of the `has_many` as definining the `where` clause and
the `has_many through:` as defining the `inner join` clause in our SQL.